fix size_t

+++ maksimov.leonid/P3/comparison.cpp
@@ -5,7 +5,7 @@ int arr::comparison(char* str, size_t lenStr, const char* etalonStr)
   int count = 0;
   for (size_t i = 0; i < lenStr; i++)
   {
-    for (size_t j = 0; j < sizeof(etalonStr) - 1; j++)
+    for (size_t j = 0; j < size_t(sizeof(etalonStr) - 1); j++)
     {
       if (int(str[i]) == int(etalonStr[j]) && int(str[i]) != 10 && int(str[i]) != 0)
       {
